+++ app/Http/Controllers/Dashboard/ProductsController.php
@@ -56,47 +56,49 @@ class ProductsController extends Controller
      */
     public function store(Request $request, Product $product)
     {
-        
-        $this->validate($request, [
-            'title' => 'required',
-            'category' => 'required',
-            'price' => 'required',
-            'image' => 'required|image',
-            'description' => 'required',
-        ]);
-        $image = $request->image;
-        $imageNewName = time(). $image->getClientOriginalName();
-        $image_path = 'images/products/';
-        $image->move($image_path, $imageNewName);
-       
-        if($request->visibility == 'NE')
-        {
-            $visibility = 0;
+        if(Auth::user()->isAdmin())
+        { 
+            $this->validate($request, [
+                'title' => 'required',
+                'category' => 'required',
+                'price' => 'required',
+                'image' => 'required|image',
+                'description' => 'required',
+            ]);
+            $image = $request->image;
+            $imageNewName = time(). $image->getClientOriginalName();
+            $image_path = 'images/products/';
+            $image->move($image_path, $imageNewName);
+           
+            if($request->visibility == 'NE')
+            {
+                $visibility = 0;
+            }
+            else
+            {
+                $visibility = 1;
+            }
+            if($request->productType == 'KOM')
+            {
+                $productType = 1;
+            }
+            else
+            {
+                $productType = 0;
+            }
+            $product = Product :: create([
+                'title' => $request->title,
+                'cat_id' => $request->category,
+                'price' => $request->price,
+                'image' => $imageNewName,
+                'description' => $request->description,
+                'visibility' => $visibility,
+                'quantity' => $productType 
+            ]);
+            Session::flash('success', 'Uspešno ste dodali novi proizvod');
+            return redirect()->route('products.index');
         }
-        else
-        {
-            $visibility = 1;
-        }
-        if($request->productType == 'KOM')
-        {
-            $productType = 1;
-        }
-        else
-        {
-            $productType = 0;
-        }
-        $product = Product :: create([
-            'title' => $request->title,
-            'cat_id' => $request->category,
-            'price' => $request->price,
-            'image' => $imageNewName,
-            'description' => $request->description,
-            'visibility' => $visibility,
-            'quantity' => $productType 
-        ]);
-        Session::flash('success', 'Uspešno ste dodali novi proizvod');
-        return redirect()->route('products.index');
-        
+        return view('privileges');
     }
 
     /**
@@ -134,41 +136,44 @@ class ProductsController extends Controller
      */
     public function update(Request $request, Product $product)
     {
-        $this->validate($request, [
-            'title' => 'required',
-            'category' => 'required',
-            'price' => 'required',
-            'description' => 'required',
-        ]);
-        if($request->hasFile('image')):
-            $image = $request->image;
-            $image_path = 'images/products/';
-            $imageNewName = time() . $image->getClientOriginalName();
-            $image->move($image_path, $imageNewName);
-            $product->image = $imageNewName;
-        endif; 
-        if($request->visibility == 'NE'):
-            $visibility = 0;
-        else:
-            $visibility = 1;
-        endif;
-        if($request->productType == 'KOM'):
-            $productType = 1;
-        else:
-            $productType = 0;
-        endif;
+        if(Auth::user()->isAdmin())
+        {
+            $this->validate($request, [
+                'title' => 'required',
+                'category' => 'required',
+                'price' => 'required',
+                'description' => 'required',
+            ]);
+            if($request->hasFile('image')):
+                $image = $request->image;
+                $image_path = 'images/products/';
+                $imageNewName = time() . $image->getClientOriginalName();
+                $image->move($image_path, $imageNewName);
+                $product->image = $imageNewName;
+            endif; 
+            if($request->visibility == 'NE'):
+                $visibility = 0;
+            else:
+                $visibility = 1;
+            endif;
+            if($request->productType == 'KOM'):
+                $productType = 1;
+            else:
+                $productType = 0;
+            endif;
 
-        $product->title = $request->title;
-        $product->cat_id = $request->category;
-        $product->price = $request->price;
-        $product->description = $request->description; 
-        $product->visibility = $visibility;
-        $product->quantity = $productType;
-        //dd($request->all());
-        $product->save();
-        Session::flash('success', 'Proizvod je uspešno ažuriran');
-        return redirect()->route('products.index');
-        
+            $product->title = $request->title;
+            $product->cat_id = $request->category;
+            $product->price = $request->price;
+            $product->description = $request->description; 
+            $product->visibility = $visibility;
+            $product->quantity = $productType;
+            //dd($request->all());
+            $product->save();
+            Session::flash('success', 'Proizvod je uspešno ažuriran');
+            return redirect()->route('products.index');
+        }
+        return view('privileges');
     }
 
     /**
@@ -179,23 +184,27 @@ class ProductsController extends Controller
      */
     public function destroy(Product $product)
     {
-        //dd('images/products/' . $product->image);
-        if(!empty($product->image))
+        if(Auth::user()->isAdmin())
         { 
-           $image_path = 'images/products/'; 
-           if(file_exists($image_path . $product->image))
-           {
-                unlink($image_path . $product->image); 
-           }
-           
-           if($product->image !== NULL)
-           { 
+            //dd('images/products/' . $product->image);
+            if(!empty($product->image))
+            { 
+               $image_path = 'images/products/'; 
+               if(file_exists($image_path . $product->image))
+               {
+                    unlink($image_path . $product->image); 
+               }
+               
+               if($product->image !== NULL)
+               { 
+                    $product->delete();
+               }
+            }else{ 
                 $product->delete();
-           }
-        }else{ 
-            $product->delete();
-        } 
-        session()->flash('success', 'Proizvod je uspešno obrisan');
-        return redirect(route('products.index'));
+            } 
+            session()->flash('success', 'Proizvod je uspešno obrisan');
+            return redirect(route('products.index'));
+        }
+        return view('privileges');
     }
 }
